requirements.txt
celery
Django
django-dirtyfields
django-reversion
djangorestframework
flower
psycopg2
rdflib
sparqlwrapper
requests